Refactor location of actuator counters class
